Update README.md
A fork of "Tajpi" from PumpedSardines. This version removes any automatic updates (for people who don't feel comfortable with the application making outbound calls in the background to the internet). This version also removes Swedish language (since that will not apply to most people). This version also makes the README more clear to explain, as well as updates the dmg file for ease of use.